# doc/swagger.yaml
swagger: "2.0"
info:
  description: "This is the FlopBox swagger."
  version: "1.0.0"
  title: "Swagger FlopBox"
host: "localhost:8080"
basePath: "/flopbox"
...
schemes:
- "http"
paths:
  /account:
    post:
      tags:
      - "account"
      summary: "Creates new account"
      operationId: "createAccount"
      parameters:
      - in: "header"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: "string"
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Account created successfuly"
        "409":
          description: "Account already exists"
    get:
      tags:
      - "account"
      summary: "Retrieves all accounts"
      operationId: "getAccounts"
      produces:
      - "text/plain"
      responses:
        "200":
          description: "Accounts retrieved successfuly"
        "409":
          description: "No account found"
    put:
      tags:
      - "account"
      summary: "Modifies password"
      operationId: "updatePasswordAccount"
      parameters:
      - in: "header"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: "string"
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      - in: "header"
        name: "newpassword"
        description: "New FlopBox login password"
        required: true
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Account's password updated successfuly"
        "409":
          description: "Account doesn't exists or bad current password"
    delete:
      tags:
      - "account"
      summary: "Deletes account"
      operationId: "deleteAccount"
      parameters:
      - in: "header"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: "string"
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "Account deleted successfuly"
        "409":
          description: "Account doesn't exists or bad current password"
  /server/{username}:
    post:
      tags:
      - "server"
      summary: "Creates new FTP server"
      operationId: "createFTPServer"
      parameters:
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      - in: "header"
        name: "address"
        description: "FTP server address"
        required: true
        type: "string"
      - in: "header"
        name: "port"
        description: "FTP server port"
        required: true
        type: "integer"
      - in: "path"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "aliasName"
        description: "FTP alias"
        required: true
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "FTP server created successfuly"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "FTP server already exists"
    get:
      tags:
      - "server"
      summary: "Retrieves all FTP server"
      operationId: "getFTPServers"
      parameters:
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      - in: "path"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: string
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "FTP servers retrieved successfuly"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "No FTP server found"
    put:
      tags:
      - "server"
      summary: "Updates FTP server"
      operationId: "updateFTPServer"
      parameters:
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      - in: "path"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      - in: "query"
        name: "newAlias"
        description: "New FTP alias"
        required: false
        type: "string"
      - in: "query"
        name: "newUri"
        description: "New FTP URI"
        required: false
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "FTP server updated successfuly"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
    delete:
      tags:
      - "server"
      summary: "Deletes FTP server"
      operationId: "deleteFTPServer"
      parameters:
      - in: "header"
        name: "password"
        description: "FlopBox login password"
        required: true
        type: "string"
      - in: "path"
        name: "username"
        description: "FlopBox login username"
        required: true
        type: string
      - in: "query"
        name: "alias"
        description: "FTP alias"
        required: true
        type: "string"
      produces:
      - 'text/plain'
      responses:
        "200":
          description: "FTP server updated successfuly"
        "401":
          description: "FlopBox bad authentication"
        "409":
          description: "Param error | Alias doesn't exists"
...